Code cleaned and refactored. Added Encrypt class for MD5 and SHA256.

+++ Domains/HS.Domains.App/FileChunk.cs
@@ -3,7 +3,7 @@
     public class FileChunk
     {
         //public FileChunk(int chunkIndex, int offset, ReadOnlyMemory<byte> payload, string md5)
-        public FileChunk(int chunkIndex, int offset, byte[] payload, string md5)
+        public FileChunk(int chunkIndex, long offset, byte[] payload, string md5)
         {
             ChunkIndex = chunkIndex;
             Offset = offset;
@@ -12,8 +12,8 @@
         }
 
         public int ChunkIndex { get; set; }
-        public int Offset { get; set; }
-        //public ReadOnlyMemory<byte> Payload { get; set; }
+        public long Offset { get; set; }
+
         public byte[] Payload { get; set; }
 
         public string CheckSumMD5 { get; set; }
